Move s2_geo processor to directory with same name
